+++ src/module_bindings/add_comment_reducer.ts
@@ -11,6 +11,6 @@ import {
 } from "spacetimedb";
 
 export default {
-  blockId: __t.u32(),
+  blockId: __t.u64(),
   text: __t.string(),
 };
